Fix 404 page and change PDF button to Print
